requirements.txt
asgiref==3.2.10
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.1
django-allauth==0.42.0
django-crispy-forms==1.9.0
django-json==0.27
django-jsonfield==1.4.0
django-map-widgets==0.3.0
django-material-admin==1.7.4
django-picklefield==3.0.1
django-tinymce==3.0.2
django-uuidfield-2==0.6.6
django-widget-tweaks==1.4.8
idna==2.9
Jinja2==2.10.1
MarkupSafe==1.1.1
mysqlclient==1.4.6
oauthlib==3.1.0
Pillow==6.2.2
psycopg2==2.8.4
psycopg2-binary==2.8.5
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2018.5
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
stripe==2.50.0
twilio==6.43.0
urllib3==1.25.9
